+++ sshd-sftp/src/main/java/org/apache/sshd/sftp/client/fs/SftpFileSystemProvider.java
@@ -658,6 +658,11 @@ public class SftpFileSystemProvider extends FileSystemProvider {
                     client.close();
                 }
             }
+
+            @Override
+            public void write(byte[] b, int off, int len) throws IOException {
+                out.write(b, off, len);
+            }
         };
     }